# Refactor input iteration in Splay