Basically, work with `&[u8]` input instead of advancing through a (buffered) input file. This is an innocently looking change that has some interesting effect on the loading performance:

- Benchmark score: 10.880 iter/s -> 10.060 iter/s (slower by ~7%)
- Level app loading in debug: 8600ms -> 960ms (faster by ~9 times (!))
- Level app loading in release: 255ms -> 150ms (faster by > 40%)

TL;DR: be careful with benchmarks! the problem with it was that the benchmark loaded the whole file in memory before testing, since it needs to run multiple times. This aspect steered the actual results away from the main bottleneck of the use case: file I/O.

With this in, a sub-second loading is totally manageable during development, it's no longer a concern.

Fixes #17